requirements.txt
numpy==2.3.3
scipy==1.16.2
pandas==2.3.3
matplotlib==3.10.6
seaborn==0.13.2
scikit-learn==1.7.2
statsmodels==0.14.5
torch==2.6.0+cu126
tensorly==0.9.0
PyYAML==6.0.3
tqdm